julio

+++ src/interfacetributavel/Tributos.java
@@ -1,6 +1,6 @@
 package interfacetributavel;
 
 public interface Tributos {
-    public void taxaMnutencao(int valor);
+    public void taxaMnutencao(double valor);
     public void impostoRenda(double valor);
 }
